+++ projects/formbuilder/src/interfaces/IItemRenderProps.ts
@@ -13,7 +13,7 @@ export interface IItemRenderProps<T extends IFormItem> {
     /** the 'onBlur' event of the properties render element */
     onBlur: (property: string, value: any) => void;
     /** overrides used to hide or disable some properties in the forms, overriding the schema settings */
-    propertyOverrides?: IPropertyOverrides;
+    propertyOverrides?: IPropertyOverrides<T>;
     /** the results of the validation */
     validationResults: IValidationResult;
     /** a prefix added to the validation results. used for distincting between nested vailadtionresults */