+++ projects/formbuilder/src/interfaces/IPropertyOverrides.ts
@@ -1,6 +1,8 @@
-export interface IPropertyOverrides {
+import { IFormListColumnsPicks } from "../components/config/IFormBuilderListConfig";
+
+export interface IPropertyOverrides<T> {
     /** array of properties to disable - or 'true' for all */
     disabledProps?: Array<string> | boolean;
     /** array of properties to hide */
-    hiddenProps?: Array<string>;
+    hiddenProps?: IFormListColumnsPicks<T>;
 }